Mejora logica

+++ src/main/java/acciones/EsperarTiempo.java
@@ -7,7 +7,7 @@ public class EsperarTiempo {
     }
 
     public static void esperar(int segundos) throws InterruptedException {
-        long tiempoMiliSegundos = Duration.ofSeconds((long)segundos).toMillis();
+        long tiempoMiliSegundos = Duration.ofSeconds(segundos).toMillis();
         Thread.sleep(tiempoMiliSegundos);
     }
 }
